dev: mm: make get_free_pages return void *
and update callsites with spatch:

@@
type T;
T addr;
expression gfp, order;
@@
(
- addr = __get_free_pages(gfp, order)
+ addr = (T)__get_free_pages(gfp, order)
|
- addr = __get_free_page(gfp)
+ addr = (T)__get_free_page(gfp)
|
- addr = get_zeroed_page(gfp)
+ addr = (T)get_zeroed_page(gfp)
)

@@
type T;
@@
(
- (T *)
  \(__get_free_pages\|get_zeroed_page\|__get_free_page\)(...)
)

+++ arch/arm64/mm/mmu.c
@@ -1158,7 +1158,7 @@ static int __init __kpti_install_ng_mappings(void *__unused)
 	remap_fn = (void *)__pa_symbol(idmap_kpti_install_ng_mappings);
 
 	if (!cpu) {
-		alloc = __get_free_pages(GFP_ATOMIC | __GFP_ZERO, order);
+		alloc = (u64)__get_free_pages(GFP_ATOMIC | __GFP_ZERO, order);
 		kpti_ng_temp_pgd = (pgd_t *)(alloc + (levels - 1) * PAGE_SIZE);
 		kpti_ng_temp_alloc = kpti_ng_temp_pgd_pa = __pa(kpti_ng_temp_pgd);
